requirements.txt
asgiref==3.5.0
Django==4.0.2
django-simple-history==3.0.0
django-utils-six==2.0
django-widget-tweaks==1.4.12
Pillow==9.0.1
psycopg2-binary==2.9.3
sqlparse==0.4.2
tzdata==2021.5
